requirements.txt
datasets==2.19.0
librosa==0.10.1
soundfile==0.12.1
numpy==1.26.4
scikit-learn==1.4.2
transformers==4.41.0
faiss-cpu==1.7.4